80 commits / 271 files from the sibling account merged as a semantic union under three rules:
never weaken a gate (stricter value wins unless a documented measured correction), keep both
sides' additions, py_compile everything. 60 conflicts resolved -- 54 by a 14-agent workflow
(resolutions logged per-file), 3 by hand (validation.py, max_audit.py, pyproject.toml), plus
crontab.manifest/desk_lessons.jsonl unions.

The load-bearing resolutions:
- libs/autodiscovery/validation.py: ours' campaign_gate_stats per-candidate path + R0224
  single-deflation kept; theirs' pc_pbo/pc_p thinner seam accepted as a second per-candidate
  path (same statistic, same R0224 rule); theirs' stationarity gate grafted WITH unmeasured
  semantics instead of silent skip; universe_returns/buy_hold_returns threaded into the
  baseline gate. Their tests/validation/test_per_candidate.py passes against the union.
- scripts/max_audit.py: theirs' density-based doctrine check replaces the size check; both
  sides' gap-register machinery kept; theirs' 5 new checks registered; the two same-named
  check_constitution functions split (reach/ratchet vs presence/review -- the second was
  shadowing the first); check_model_freshness restored from theirs (lost in the merge);
  split_acked made width-preserving so evidence-scoped defect tuples survive the ack filter.
- Coverage nets caught and fixed real holes: kimi_hunter + run_strategic_director now carry
  OBJECTIVE_PREAMBLE; three measurement scripts + quant-alerts/cadence/daily-max timers wired
  into the scheduler manifest; 19 new fences claimed under governing laws in the enforcement
  matrix; 4 new docs classified in artifact governance.
- Lessons ledger union: 7 colliding ids renumbered with provenance; curation debt paid by
  graduation not budget -- L0052 (browser-UA) and L0038/L0048/L0054 (mined-evidence priority,
  now in every miner prompt) pinned by new tests; L0061 graduated by BUILDING its fix
  (okx_inst 1000-prefix re-denomination: 260/653 -> full resolution, SHIB/PEPE/FLOKI/BONK/SATS
  recovered); L0024 retired as absorbed into L0038.
- Push-ladder standing policy extended to the CRO itself: run_cro now uses push_rounds via a
  new llm_seat.chat_messages (same cap-before-call/degrade/spend machinery), rounds parsed
  separately with the duplicate fence deduping across rounds.
- DATA_UNIVERSE_TAXONOMY.md installed (principal 2026-08-04): 30-domain hunting map + bulk
  name-list, governed under section 36 weekly, injected whole into breadth_expander's pushed
  system prompt.
- Defects fixed in passing: research_alpha_optimizer f-string called as function (runtime
  TypeError); naive datetime in conversion_max; stale test pinning pre-R0080 capacity
  behaviour; six_wirings source-phrase probes upgraded to behavioural tests; jsonl integrity
  fence taught the lessons ledger's own '#' header convention.

Verification: ruff clean, mypy clean (480 files), law gate 10/10 PASS, full pytest green
(3 environment skips: arch, vectorbt, no-urgent-page).

+++ libs/research/evidence_clock.py
@@ -86,11 +86,12 @@ def sufficient(mean: float, stdev: float, n_obs: int, *,
                            min_obs - n_obs)
     if t < min_t:
         # n needed scales with the SQUARE of the t shortfall, because t grows as sqrt(n).
-        need = int(math.ceil(n_obs * (min_t / t) ** 2)) if t > 0 else 0
+        need = math.ceil(n_obs * (min_t / t) ** 2) if t > 0 else 0
         short = max(0, need - n_obs) if need else -1
         return Sufficiency(False, t, n_obs,
                            f"t={t:.2f} < {min_t} at n={n_obs}" +
                            (f" -- needs ~{short} more observations at this effect size"
-                            if short >= 0 else " -- effect is non-positive, more data will not fix it"),
+                            if short >= 0
+                            else " -- effect is non-positive, more data will not fix it"),
                            short if short >= 0 else 0)
     return Sufficiency(True, t, n_obs, f"t={t:.2f} >= {min_t} on {n_obs} observations", 0)
